requirements.txt
streamlit>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.0
pypdf>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
pandas>=2.0.0
youtube-transcript-api>=0.6.0
beautifulsoup4>=4.12.0
requests>=2.31.0
plotly>=5.18.0
